requirements.txt
numpy==1.24.4
pandas==1.5.3
seaborn==0.13.2
nltk==3.8.1
requests==2.31.0
scikit-learn==1.2.2
PyYAML==6.0.1
mlflow==2.10.0
fastapi==0.109.0
uvicorn==0.27.0
ipykernel==6.29.0
pycaret==3.2.0